DRY up what all 3 classes have in common
