+++ apprentice_teacher.rb
@@ -1,16 +1,12 @@
-class ApprenticeTeacher
-  attr_reader :age, :salary, :phase, :target_raise
-  attr_accessor :name
+require_relative 'dbc_member'
+
+class ApprenticeTeacher < DBCMember
+  attr_reader :salary, :target_raise
 
   def initialize(options={})
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
+    super
     @target_raise = 800
     @phase = 3
-  end
-
-  def offer_high_five
-    "High five!"
   end
 
   def set_phase(num)
