+++ high_fiveable.rb
@@ -1,0 +1,5 @@
+module HighFiveable
+  def offer_high_five
+    "High five!"
+  end
+end
