+++ senior_teacher.rb
@@ -1,16 +1,12 @@
-class SeniorTeacher
-  attr_reader :age, :salary, :phase, :performance_rating, :target_raise
-  attr_accessor :name
+require_relative 'dbc_member'
+
+class SeniorTeacher < DBCMember
+  attr_reader :salary, :performance_rating, :target_raise
 
   def initialize(options={})
+    super
     @phase = 3
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
     @target_raise = 1000
-  end
-
-  def offer_high_five
-    "High five!"
   end
 
   def set_phase(num)
